+++ src/integrationTest/java/org/manuel/mysportfolio/controllers/query/CompetitionQueryControllerTest.java
@@ -1,6 +1,7 @@
 package org.manuel.mysportfolio.controllers.query;
 
 import static java.time.DayOfWeek.MONDAY;
+import static java.time.DayOfWeek.SATURDAY;
 import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
 import static org.manuel.mysportfolio.model.Sport.FOOTBALL;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
@@ -8,6 +9,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import java.time.YearMonth;
 import javax.inject.Inject;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.AfterEach;
@@ -56,6 +58,26 @@ public class CompetitionQueryControllerTest {
         .andExpect(status().isOk())
         .andExpect(jsonPath("$.content", hasSize((int) competitionRepository.count())))
         .andExpect(jsonPath("$.content[0].name", Matchers.is(competitionSaved.getName())));
+  }
+
+  @Test
+  public void testGetActiveCompetitions() throws Exception {
+    final var competitionSaved = competitionRepository.save(
+        new Competition(null, null, "KNVB 4th Klass 2019", FOOTBALL, MONDAY,
+            YearMonth.of(2019, 1), YearMonth.of(2020, 1), null));
+
+    final var competitionSaved2 = competitionRepository.save(
+        new Competition(null, null, "KNVB 4e Klass 2020", FOOTBALL, SATURDAY,
+            YearMonth.of(2020, 1), YearMonth.of(2021, 1), null));
+    final var competitionSaved3 = competitionRepository.save(
+        new Competition(null, null, "KNVB 1e Klass 2020", FOOTBALL, SATURDAY,
+            YearMonth.of(2020, 1), null, null));
+
+    mvc.perform(get("/api/v1/competitions?matchDate=2020-02-01").contentType(APPLICATION_JSON))
+        .andExpect(status().isOk())
+        .andExpect(jsonPath("$.content", hasSize(2)))
+        .andExpect(jsonPath("$.content[0].name", Matchers.is(competitionSaved2.getName())))
+        .andExpect(jsonPath("$.content[1].name", Matchers.is(competitionSaved3.getName())));
   }
 
 }