+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added new badges push notification
+- Updating Spring Boot
+- Adding matchDate to competitions query
+- Adding v-shape kit
 
 ### Changed